fix: resolved issues with styling on the ride page and also home page for mobile and desktop
