+++ src/common/layouts/default/header/header.scss
@@ -1,0 +1,61 @@
+.logo {
+   font-weight: 600;
+   a {
+      font-size: 20px;
+   }
+}
+
+header {
+   display: flex;
+   justify-content: space-between;
+   align-items: center;
+   padding: 30px;
+   width: 100%;
+   max-width: 1800px;
+   margin: 0 auto;
+
+   &.mobile-overlay {
+      @media (max-width: 600px) {
+         position: fixed;
+         z-index: 1;
+         top: 0;
+         left: 50%;
+         transform: translateX(-50%);
+
+         .get-started {
+            display: none;
+         }
+      }
+   }
+}
+
+.nav-links {
+   display: flex;
+   align-items: center;
+   justify-content: space-between;
+   width: 400px;
+
+   @media screen and (max-width: 760px) {
+      display: none;
+   }
+}
+
+nav {
+   font-weight: 600;
+}
+
+.action-links {
+   display: flex;
+   color: white;
+   align-items: center;
+
+   & > a:first-child {
+      margin-right: 30px;
+   }
+}
+
+.get-started {
+   padding: 15px 25px;
+   background-color: #0d9648;
+   font-weight: 600;
+}
