+++ src/common/sections/home/Hero/Hero.scss
@@ -337,13 +337,13 @@
                @media screen and (min-width: 320px) {
                   overflow-y: scroll;
                   -webkit-box-pack: start;
-                  justify-content: start;
+                  justify-content: space-evenly;
                }
 
                @media screen and (min-width: 600px) {
                   overflow-y: scroll;
                   -webkit-box-pack: start;
-                  justify-content: start;
+                  justify-content: space-evenly;
                }
 
                @media screen and (min-width: 1136px) {
@@ -371,6 +371,10 @@
                   outline-offset: -2px;
                   color: #545454;
                   outline: none;
+               }
+
+               &.active {
+                  border-bottom: 3px solid black;
                }
 
                .tab__content--container {
@@ -424,15 +428,7 @@
                   display: block;
 
                   @media screen and (min-width: 320px) {
-                     min-height: 48px;
-                  }
-
-                  @media screen and (min-width: 600px) {
-                     min-height: 48px;
-                  }
-
-                  @media screen and (min-width: 1136px) {
-                     min-height: 40px;
+                     min-height: 30px;
                   }
                }
             }
