+++ src/common/sections/home/Hero/TabHeader.tsx
@@ -7,9 +7,17 @@ import { MdRestaurant } from "react-icons/md";
 interface TabHeaderItemProps {
    Icon: IconType;
    title: string;
+   index: number;
+   onClick: () => void;
+   isActive: boolean;
 }
 
-const TabHeader = () => {
+interface TabHeaderProps {
+   activeTab: number;
+   setActiveTab: React.Dispatch<React.SetStateAction<number>>;
+}
+
+const TabHeader: React.FC<TabHeaderProps> = ({ activeTab, setActiveTab }) => {
    const tabHeaders = [
       { title: "Ride", icon: AiFillCar },
       { title: "Drive", icon: GiNetworkBars },
@@ -18,20 +26,28 @@ const TabHeader = () => {
    return (
       <div className="hero__inner-content--container__tabslist" role="tablist">
          {tabHeaders.map(({ icon, title }, index) => (
-            <TabHeaderItem Icon={icon} title={title} key={index} />
+            <TabHeaderItem
+               Icon={icon}
+               title={title}
+               key={index}
+               index={index}
+               onClick={() => setActiveTab(index)}
+               isActive={activeTab === index}
+            />
          ))}
       </div>
    );
 };
 
-const TabHeaderItem: React.FC<TabHeaderItemProps> = ({ Icon, title }) => {
+const TabHeaderItem: React.FC<TabHeaderItemProps> = ({ Icon, title, index, onClick, isActive }) => {
    return (
       <div
          role="tab"
-         tabIndex={0}
+         tabIndex={index}
          id="0"
          aria-selected={false}
-         className="hero__inner-content--container__tab"
+         className={`hero__inner-content--container__tab${isActive ? " active" : ""}`}
+         onClick={onClick}
       >
          <div className="tab__content--container">
             <div>
