+++ src/common/sections/ride/ride-details/ride-details.scss
@@ -1,0 +1,106 @@
+.ride-details {
+   min-height: 100vh;
+   bottom: auto;
+   position: relative;
+   top: 44vh;
+   width: 100%;
+   z-index: 2;
+   background-color: white;
+
+   @media (min-width: 600px) {
+      bottom: 16px;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      min-height: auto;
+      position: absolute;
+      top: 16px;
+      width: 406px;
+      left: 16px;
+      padding: 20px 30px;
+   }
+
+   @media (min-width: 1280px) {
+      left: calc((100vw - 1280px) / 2);
+   }
+
+   &--top {
+      padding: 16px;
+      position: relative;
+      // text-align: center;
+   }
+
+   &--drag__icon {
+      @media (max-width: 600px) {
+         background-color: rgb(226, 226, 226);
+         display: block;
+         height: 4px;
+         left: 44%;
+         position: absolute;
+         top: 8px;
+         width: 12%;
+      }
+   }
+
+   &--text {
+      padding: 16px 0;
+      position: relative;
+      color: rgb(0, 0, 0);
+      display: none;
+      font-size: 18px;
+      font-weight: 500;
+      line-height: 24px;
+      box-sizing: border-box;
+
+      @media screen and (min-width: 600px) {
+         display: block;
+         font-size: 30px;
+         font-weight: 500;
+         line-height: 44px;
+      }
+   }
+
+   &--middle {
+      padding: 0 16px;
+   }
+
+   .input-fields {
+      position: relative;
+
+      input:focus {
+         border: 2px solid black;
+      }
+
+      &--icons {
+         left: 22px;
+         bottom: 33px;
+         top: 33px;
+         background-color: #000000;
+         position: absolute;
+         width: 1px;
+
+         &::before {
+            background-color: #f6f6f6;
+            border: 1px solid #000000;
+            bottom: -10px;
+            height: 5px;
+            width: 5px;
+            left: -3px;
+            position: absolute;
+            content: "";
+         }
+
+         &::after {
+            background-color: #f6f6f6;
+            border: 1px solid #000000;
+            border-radius: 50%;
+            top: -10px;
+            height: 6px;
+            width: 6px;
+            left: -3px;
+            position: absolute;
+            content: "";
+         }
+      }
+   }
+}
