+++ src/pages/ride/index.tsx
@@ -1,13 +1,14 @@
 import DefaultLayout from "../../common/layouts/default";
 import Maps from "../../common/sections/ride/map/Map";
 import RideDetails from "../../common/sections/ride/ride-details/ride-details";
+import "./ride.scss";
 
 const Ride = () => {
    return (
-      <DefaultLayout>
+      <DefaultLayout hasMobileOverlayNav>
          <div className="ride">
-            <RideDetails />
             <Maps />
+            <RideDetails />
          </div>
       </DefaultLayout>
    );
