+++ src/pages/ride/ride.scss
@@ -1,0 +1,9 @@
+.ride {
+   height: 100%;
+   // pointer-events: none;
+   position: relative;
+
+   @media (min-width: 600px) {
+      height: calc(100vh - 111px);
+   }
+}
