+++ src/styles/index.scss
@@ -1,4 +1,3 @@
 @import "./sections/home/home.scss";
-@import "./sections/ride/ride.scss";
 
 @import "./layouts/layouts.scss";
